Using googleVis to visualize the stock markets data
This is written in R using googleVis package.Please set the database host & database name before running the script.This will ask for the Username,password to connect to database. Along with the supply the Stock quote as an input. e.g: AAPL for Apple Inc. Refer pdf file to preview the visualization. You can write to me in case any problems with the code execution.